fixing FVCalc

+++ FVCalc.java
@@ -7,7 +7,7 @@ public class FVCalc {
 		int years = Integer.parseInt(args[2]);
 		
 		// calculating future value using the formula FV = PV * (1 + r)^n
-		double futureValue = currentValue * Math.pow((1 + interestRate/100), years);	
+		double futureValue = currentValue * Math.pow((1 + interestRate/100), (double) years);	
 
 		System.out.println("After " + years + " years, " + "$" + (int) currentValue + 
 							" saved at " + interestRate + "% will yield $" + (int) futureValue);
